abstract messy functions
